💡 refactor >>storybook에 globalstyle적용
